Implement part 2: 1. Calculate the Big-O running time of your code to compute the Flesch score for BasicDocument and EfficientDocument (including the time taken by processText!) and predict how the running time of calling fleschScore on  BasicDocument and EfficientDocument will grow as the document size grows.
2. Open the file DocumentBenchmarking.java, which you can find in the document package.  You will be adding code to the main method, and you will find the method getStringFromFile useful for reading a specified number of characters from a text file.

3. Complete the main method so that it prints a table with one column each for the amount of time it takes to create a BasicDocument or an EfficientDocument, respectively, and call fleschScore on it.  Your table should look like the following:

NumberOfChars        BasicTime              EfficientTime

5000                         [basic time 1]          [efficient time 1]

10000                       [basic time 2]          [efficient time 2]

15000                       [basic time 3]          [efficient time 3]

...

Where the first column in the table prints the number of characters in the test, the second column is the amount of time it takes to create a BasicDocument with that many characters and call fleschScore on it, and the third column is the amount of time it takes to create an EfficientDocument with that many characters and call fleschScore on it.  In these timing calculations, you should include the time taken by both constructors (i.e. the time EfficientDocument takes
 calculating the number of words, syllables and sentences in the
document), otherwise it's not a fair
comparison at all!

Columns in the table should be separated with a \t character for easy copy and paste in the next section.  You can also write this table to a file if you want, but it's not necessary.

4. Plot your results and compare them to your predictions.  Copy the output of your program and paste it into a Google spreadsheet (sheets.google.com).  It should automatically paste into 3 columns.  Then, graph the data using the Explore button in the lower right corner